+++ packages/backend/src/backend_factory.test.ts
@@ -10,15 +10,23 @@ import {
   BackendDeploymentType,
   CDKContextKey,
 } from '@aws-amplify/platform-core';
+import { BackendEnvironmentVariables } from './environment_variables.js';
 
-const createStackAndSetContext = (): Stack => {
+const createStackAndSetContext = (
+  deploymentType: BackendDeploymentType
+): Stack => {
   const app = new App();
-  app.node.setContext('branch-name', 'testEnvName');
-  app.node.setContext('backend-id', 'testBackendId');
-  app.node.setContext(
-    CDKContextKey.DEPLOYMENT_TYPE,
-    BackendDeploymentType.BRANCH
-  );
+  app.node.setContext(CDKContextKey.DEPLOYMENT_TYPE, deploymentType);
+  switch (deploymentType) {
+    case BackendDeploymentType.SANDBOX:
+      app.node.setContext('backend-id', 'app-user');
+      break;
+    case BackendDeploymentType.BRANCH:
+      app.node.setContext('branch-name', 'testEnvName');
+      app.node.setContext('backend-id', 'testBackendId');
+      break;
+  }
+
   const stack = new Stack(app);
   return stack;
 };
@@ -26,7 +34,7 @@ const createStackAndSetContext = (): Stack => {
 void describe('Backend', () => {
   let rootStack: Stack;
   beforeEach(() => {
-    rootStack = createStackAndSetContext();
+    rootStack = createStackAndSetContext(BackendDeploymentType.BRANCH);
   });
 
   void it('initializes constructs in given app', () => {
@@ -134,10 +142,41 @@ void describe('Backend', () => {
     );
   });
 
-  void describe('getOrCreateStack', () => {
+  void it('registers branch linker for branch deployments if enabled', () => {
+    try {
+      process.env[
+        BackendEnvironmentVariables.AMPLIFY_BACKEND_BRANCH_LINKER_ENABLED
+      ] = 'true';
+      new BackendFactory({}, rootStack);
+      const rootStackTemplate = Template.fromStack(rootStack);
+      rootStackTemplate.resourceCountIs(
+        'Custom::AmplifyBranchLinkerResource',
+        1
+      );
+    } finally {
+      delete process.env[
+        BackendEnvironmentVariables.AMPLIFY_BACKEND_BRANCH_LINKER_ENABLED
+      ];
+    }
+  });
+
+  void it('does not register branch linker for branch deployments by default', () => {
+    new BackendFactory({}, rootStack);
+    const rootStackTemplate = Template.fromStack(rootStack);
+    rootStackTemplate.resourceCountIs('Custom::AmplifyBranchLinkerResource', 0);
+  });
+
+  void it('does not register branch linker for sandbox deployments', () => {
+    const rootStack = createStackAndSetContext(BackendDeploymentType.SANDBOX);
+    new BackendFactory({}, rootStack);
+    const rootStackTemplate = Template.fromStack(rootStack);
+    rootStackTemplate.resourceCountIs('Custom::AmplifyBranchLinkerResource', 0);
+  });
+
+  void describe('getStack', () => {
     void it('returns nested stack', () => {
       const backend = new BackendFactory({}, rootStack);
-      const testStack = backend.getOrCreateStack('testStack');
+      const testStack = backend.getStack('testStack');
       assert.strictEqual(rootStack.node.findChild('testStack'), testStack);
     });
   });