+++ packages/backend/src/backend_factory.ts
@@ -15,11 +15,14 @@ import { createDefaultStack } from './default_stack_factory.js';
 import { getUniqueBackendIdentifier } from './backend_identifier.js';
 import {
   BackendDeploymentType,
+  BranchBackendIdentifier,
   SandboxBackendIdentifier,
 } from '@aws-amplify/platform-core';
-import { stackOutputKey } from '@aws-amplify/backend-output-schemas';
+import { platformOutputKey } from '@aws-amplify/backend-output-schemas';
 import { fileURLToPath } from 'url';
 import { Backend } from './backend.js';
+import { AmplifyBranchLinkerConstruct } from './engine/branch-linker/branch_linker_construct.js';
+import { BackendEnvironmentVariables } from './environment_variables.js';
 
 // Be very careful editing this value. It is the value used in the BI metrics to attribute stacks as Amplify root stacks
 const rootStackTypeIdentifier = 'root';
@@ -49,7 +52,10 @@ export class BackendFactory<
       rootStackTypeIdentifier,
       fileURLToPath(new URL('../package.json', import.meta.url))
     );
-    this.stackResolver = new NestedStackResolver(stack);
+    this.stackResolver = new NestedStackResolver(
+      stack,
+      new AttributionMetadataStorage()
+    );
 
     const constructContainer = new SingletonConstructContainer(
       this.stackResolver
@@ -60,15 +66,25 @@ export class BackendFactory<
     );
 
     const uniqueBackendIdentifier = getUniqueBackendIdentifier(stack);
-    outputStorageStrategy.addBackendOutputEntry(stackOutputKey, {
+    outputStorageStrategy.addBackendOutputEntry(platformOutputKey, {
       version: '1',
       payload: {
         deploymentType:
           uniqueBackendIdentifier instanceof SandboxBackendIdentifier
             ? BackendDeploymentType.SANDBOX
             : BackendDeploymentType.BRANCH,
+        region: stack.region,
       },
     });
+
+    const shouldEnableBranchLinker =
+      uniqueBackendIdentifier instanceof BranchBackendIdentifier &&
+      process.env[
+        BackendEnvironmentVariables.AMPLIFY_BACKEND_BRANCH_LINKER_ENABLED
+      ] === 'true';
+    if (shouldEnableBranchLinker) {
+      new AmplifyBranchLinkerConstruct(stack, uniqueBackendIdentifier);
+    }
 
     const importPathVerifier = new ToggleableImportPathVerifier();
 
@@ -99,10 +115,11 @@ export class BackendFactory<
   }
 
   /**
-   * Returns a CDK stack within the Amplify project that can be used for creating custom resources
+   * Returns a CDK stack within the Amplify project that can be used for creating custom resources.
+   * @returns existing stack if provided name has been used or create new one with the provided name
    */
-  getOrCreateStack = (name: string): Stack => {
-    return this.stackResolver.getStackFor(name);
+  getStack = (name: string): Stack => {
+    return this.stackResolver.getCustomStack(name);
   };
 }