+++ packages/backend/src/engine/singleton_construct_container.test.ts
@@ -6,11 +6,16 @@ import assert from 'node:assert';
 import { Construct } from 'constructs';
 import { Queue } from 'aws-cdk-lib/aws-sqs';
 import { NestedStackResolver } from './nested_stack_resolver.js';
-import { ConstructContainerEntryGenerator } from '@aws-amplify/plugin-types';
+import {
+  ConstructContainer,
+  ConstructContainerEntryGenerator,
+  ConstructFactory,
+} from '@aws-amplify/plugin-types';
 import {
   BackendDeploymentType,
   CDKContextKey,
 } from '@aws-amplify/platform-core';
+import { AttributionMetadataStorage } from '@aws-amplify/backend-output-storage';
 
 const createStackAndSetContext = (): Stack => {
   const app = new App();
@@ -32,7 +37,7 @@ void describe('SingletonConstructContainer', () => {
     });
     void it('calls initializer to create construct instance', () => {
       const container = new SingletonConstructContainer(
-        new NestedStackResolver(stack)
+        new NestedStackResolver(stack, new AttributionMetadataStorage())
       );
       const instance = container.getOrCompute({
         resourceGroupName: 'testGroup',
@@ -45,7 +50,7 @@ void describe('SingletonConstructContainer', () => {
 
     void it('returns cached instance if initializer has been seen before', () => {
       const container = new SingletonConstructContainer(
-        new NestedStackResolver(stack)
+        new NestedStackResolver(stack, new AttributionMetadataStorage())
       );
       const initializer: ConstructContainerEntryGenerator = {
         resourceGroupName: 'testGroup',
@@ -61,7 +66,7 @@ void describe('SingletonConstructContainer', () => {
 
     void it('returns correct cached value for each initializer', () => {
       const container = new SingletonConstructContainer(
-        new NestedStackResolver(stack)
+        new NestedStackResolver(stack, new AttributionMetadataStorage())
       );
       const bucketInitializer: ConstructContainerEntryGenerator = {
         resourceGroupName: 'testGroup',
@@ -86,6 +91,38 @@ void describe('SingletonConstructContainer', () => {
 
       assert.strictEqual(bucket, cachedBucket);
       assert.strictEqual(queue, cachedQueue);
+    });
+  });
+
+  void describe('getConstructFactory', () => {
+    let container: ConstructContainer;
+    const testFactoryToken = 'factory1';
+    const testFactory: ConstructFactory = {
+      name: 'factory1',
+    } as unknown as ConstructFactory;
+
+    beforeEach(() => {
+      container = new SingletonConstructContainer(
+        new NestedStackResolver(
+          createStackAndSetContext(),
+          new AttributionMetadataStorage()
+        )
+      );
+    });
+
+    void it('returns for registered factory', () => {
+      container.registerConstructFactory(testFactoryToken, testFactory);
+      assert.deepStrictEqual(
+        container.getConstructFactory(testFactoryToken),
+        testFactory
+      );
+    });
+
+    void it('returns undefined for unregistered factory', () => {
+      assert.deepStrictEqual(
+        container.getConstructFactory(testFactoryToken),
+        undefined
+      );
     });
   });
 });