+++ packages/integration-tests/src/test-e2e/test_project_base.ts
@@ -60,7 +60,7 @@ export abstract class TestProjectBase {
         ],
         this.projectDirPath,
         {
-          env: { CI: 'true' },
+          env: { CI: 'true', AMPLIFY_BACKEND_BRANCH_LINKER_ENABLED: 'true' },
         }
       ).run();
     }