+++ packages/backend-graphql/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @aws-amplify/backend-graphql
 
+## 0.3.2
+
+### Patch Changes
+
+- 3bff764b: Expose user pool and client as cfn resources.
+- Updated dependencies [3bff764b]
+  - @aws-amplify/plugin-types@0.3.1
+
+## 0.3.1
+
+### Patch Changes
+
+- 0bd8a3f3: add missing dev deps
+
+## 0.3.0
+
+### Minor Changes
+
+- 42127d0a: Add support for functions as input to the defineData call.
+
+### Patch Changes
+
+- 1dd824cb: Bump graphql-api construct version
+
+## 0.2.1
+
+### Patch Changes
+
+- 457b1662: getConstructFactory no longer throws an error if the factory is not found, and returns undefined instead.
+- 46e0aad6: Update backend-graphql library to start decomposing translation layer from backend definitions into CDK
+- Updated dependencies [79a6e09f]
+- Updated dependencies [457b1662]
+- Updated dependencies [79a6e09f]
+  - @aws-amplify/backend-output-schemas@0.2.1
+  - @aws-amplify/plugin-types@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes