+++ packages/deployed-backend-client/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @aws-amplify/deployed-backend-client
 
+## 0.2.2
+
+### Patch Changes
+
+- bbb5b657: update sdk client deps
+- 1cdb3ba9: add DELETING backend deployment status
+
+## 0.2.1
+
+### Patch Changes
+
+- 79a6e09f: Change stackOutputKey to platformOutputKey
+- Updated dependencies [79a6e09f]
+- Updated dependencies [79a6e09f]
+  - @aws-amplify/backend-output-schemas@0.2.1
+  - @aws-amplify/platform-core@0.1.2
+
 ## 0.2.0
 
 ### Minor Changes