+++ packages/integration-tests/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @aws-amplify/integration-tests
 
+## 0.2.2
+
+### Patch Changes
+
+- 79a6e09f: Change stackOutputKey to platformOutputKey
+- 93c956ea: Add aws_project_region to e2e test assert
+- 79a6e09f: Add aws_project_region to amplifyconfiguration.json
+
+## 0.2.1
+
+### Patch Changes
+
+- d0119b25: Flatten loginWith type to improve autocompletion.
+- 4e48e4ba: chore: add new defineBackend to better align with other backend factories
+
 ## 0.2.0
 
 ### Minor Changes