+++ packages/platform-core/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @aws-amplify/platform-core
 
+## 0.1.3
+
+### Patch Changes
+
+- 0bd8a3f3: add 'main' entry to package.json
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [457b1662]
+  - @aws-amplify/plugin-types@0.3.0
+
 ## 0.1.1
 
 ### Patch Changes